+++ BuffCrops/release-notes.md
@@ -1,5 +1,13 @@
 # Buff Crops Changelog
 
+## 1.0.2
+#### Additions
+* Added a config to enable debug logging.
+
+## 1.0.1
+#### Fixes
+* Crops should now only contribute to buffs when fully grown as they were originally meant to. Oops!
+
 ## 1.0.0
 #### Additions
 * The mod exists!